+++ test/support/domain.ex
@@ -23,5 +23,6 @@ defmodule AshWorkflowTest.Domain do
     resource AshWorkflowTest.ErrorPathWorkflow
     resource AshWorkflowTest.CheckIntervalWorkflow
     resource AshWorkflowTest.AuthorizedErrorPathWorkflow
+    resource AshWorkflowTest.SharedTimeoutNameWorkflow
   end
 end